Separated Downloader class to the designated file
