+++ CMakeLists.txt
@@ -10,7 +10,8 @@ set_property(TARGET tdjson_example PROPERTY CXX_STANDARD 11)
 
 include_directories(inc)
 add_executable(td_downloader ./src/td_downloader.cpp)
-add_library(TaskApi ./src/task_api.cpp)
+add_library(TaskApi ./src/task_api.cpp
+                    ./src/Downloader.cpp)
 target_link_libraries(TaskApi PRIVATE Td::TdStatic)
 target_link_libraries(td_downloader PRIVATE Td::TdStatic TaskApi)
 set_property(TARGET td_downloader PROPERTY CXX_STANDARD 14)
